+++ project_electron/orgs/admin.py
@@ -4,44 +4,89 @@ from __future__ import unicode_literals
 from django.contrib import admin
 from orgs.models import Organization, User, BAGLogCodes, BAGLog, Archives, BagInfoMetadata, LanguageCode, BagItProfile, BagItProfileBagInfo
 from orgs.accession.models import Accession
+from orgs.rights.models import *
 
 
 @admin.register(Organization)
 class OrganizationsAdmin(admin.ModelAdmin):
-	pass
+    pass
+
 
 @admin.register(User)
 class UserAdmin(admin.ModelAdmin):
-	pass
+    pass
+
 
 @admin.register(BAGLogCodes)
 class BagCodesAdmin(admin.ModelAdmin):
-	pass
+    pass
+
 
 @admin.register(BAGLog)
 class BagLogAdmin(admin.ModelAdmin):
-	pass
+    pass
+
 
 @admin.register(Archives)
 class ArchivesAdmin(admin.ModelAdmin):
-	pass
+    pass
+
 
 @admin.register(BagInfoMetadata)
 class BagInfoMetadata(admin.ModelAdmin):
-	pass
+    pass
+
 
 @admin.register(Accession)
 class AccessionAdmin(admin.ModelAdmin):
     pass
 
+
 @admin.register(BagItProfile)
 class BagItProfile(admin.ModelAdmin):
-	pass
+    pass
+
 
 @admin.register(BagItProfileBagInfo)
 class BagItProfileBagInfo(admin.ModelAdmin):
-	pass
+    pass
+
 
 @admin.register(LanguageCode)
 class BagInfoMetadata(admin.ModelAdmin):
-	pass
+    pass
+
+
+@admin.register(RightsStatement)
+class RightsStatementAdmin(admin.ModelAdmin):
+    pass
+
+
+@admin.register(RightsStatementCopyright)
+class RightsStatementCopyrightAdmin(admin.ModelAdmin):
+    pass
+
+
+@admin.register(RightsStatementLicense)
+class RightsStatementLicenseAdmin(admin.ModelAdmin):
+    pass
+
+
+@admin.register(RightsStatementStatute)
+class RightsStatementStatuteAdmin(admin.ModelAdmin):
+    pass
+
+
+@admin.register(RightsStatementOther)
+class RightsStatementOtherAdmin(admin.ModelAdmin):
+    pass
+
+
+@admin.register(RightsStatementRightsGranted)
+class RightsStatementRightsGrantedAdmin(admin.ModelAdmin):
+    pass
+
+
+@admin.register(RecordType)
+class RecordType(admin.ModelAdmin):
+    pass
