+++ project_electron/orgs/api/serializers.py
@@ -3,7 +3,7 @@ from __future__ import unicode_literals
 
 from rest_framework import serializers
 from orgs.models import *
-from rights.models import *
+from orgs.rights.models import *
 
 
 class RightsStatementRightsGrantedSerializer(serializers.ModelSerializer):
