+++ project_electron/orgs/templates/rights/modals/confirm_delete.html
