+++ project_electron/project_electron/settings.py
@@ -44,7 +44,6 @@ INSTALLED_APPS = [
     'rac_user',
     'transfer_app',
     'orgs',
-    'rights',
 ]
 
 MIDDLEWARE = [
