+++ project_electron/transfer_app/static/js/delete_object_js.js
